Changes to Linear
